Use pattern matching to catch typos.
Being able to use any placeholder adds to the magical feel of Methodsolver, but it comes at the cost of silently failing because typos. The plan is to limit placeholders to names that start with a blank. And, to foo bar and friends.

Also, added specs for the safety features.